% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expand.R
\name{data_expand}
\alias{data_expand}
\title{Create an expanded data frame}
\usage{
data_expand(
  data,
  id.var,
  outcome.var,
  time.var,
  cov.ti,
  cov.td,
  int.width,
  bandwidth,
  width.subset = FALSE
)
}
\arguments{
\item{data}{is a data frame in a long format}

\item{id.var}{is the variable name indicating ID or similar}

\item{outcome.var}{is the variable name of the outcome}

\item{time.var}{is the variable name of the follow-up time}

\item{cov.ti}{is the list of variable names of the time-independent covariates}

\item{cov.td}{is the list of variable names of the time-dependent covariates}

\item{int.width}{is the interval width}

\item{bandwidth}{is the bandwidth}

\item{width.subset}{is a logistic argument that determines whether or not to
subset the analysis data: \code{TRUE}, subsetting the data based on \code{int.width}
and \code{bandwidth}; \code{FALSE}, otherwise.}
}
\value{
an expanded data frame in long format
}
\description{
Function short description
}
\references{
Neal, J. E., Zhang, P. & Liu, D. (2024). Predictive partly conditional
models for longitudinal ordinal outcomes with application to Alzheimer's disease
progression. Under review.
}
\keyword{internal}
